timestamps: stamp on history add, not send
Two related changes:
1. Timestamps are added to each cache change when written to a writer's
history cache. Previously, the timestamp was taken when the message was
being sent on the wire, so a retry would result in the timestamp being
updated on the resend.
2. Multiple INFO_TS submessages are sent, one for each cache change
being sent, with that cache change's timestamp. Previously, a single
INFO_TS message was inserted (with the current time), followed by
(possibly) multiple cache changes. Inserting a timestamp before each
cache change is needed to keep track of each message's timestamp.

I could not find anything in the RTPS spec that indicates whether the
old way or this way is correct. It *seems* like the intent of the
INFO_TS submessage, particularly the 'invalidate' flag, is that it be
used multiple times in a message (as is done here now).

Also, the timestamp associated with a cache change uses the monotonic
clock. This should be configurable somehow; some systems will certainly
want to use wall-clock time, whereas others (like ours) do not have
wall-clock time. (In any case, the 'interval' functions in eClock should
use the monotonic clock.)

A final detail: when a cache change is initialized (reset), the source
timestamp is now set to "invalid" instead of zero. This allows the
generation of the INFO_TS submessages to correctly set the 'invalid'
flag in the submessage, which is what makes it so we can tell whether
a message on the receiving end has a valid timestamp or not.