network: all udp sends from 0.0.0.0
Previously, the stack was queried and every packet was sent from each
of the host's IP addresses. This change makes it so there is one send
socket bound to 0.0.0.0, instead of one send socket bound to each host
IP.

Packets are still sent multiple times. For example, if a host has IPs
127.0.0.1 and 192.168.1.100, and a writer is matched to a reader on
port 10040, each packet will be sent to 127.0.0.1:10040 and to
192.168.1.100:10040.

This reduces network traffic by its square root, but there is still room
for improvement. If there are N host IPs, traffic is reduced from O(N^2)
to O(N), but needs to be reduced to O(1).