test: add history tests
These tests exercise various history edge cases, mainly overruns, on
both the publisher and subscriber sides. Main test loop looks like
this:

for (ReliabilityQosPolicyKind qos : {BEST_EFFORT_RELIABILITY_QOS, RELIABLE_RELIABILITY_QOS})
  for (int pubHist : {histSmall, histLarge})
    for (int subHist : {histSmall, histLarge})

where "histSmall" is a history of depth that we can easily overrun it,
and "histLarge" is of depth that the test won't overrun it.