history: don't block reliable writer
This may be a controversial change. The idea is that if a publisher has
reliable connections to multiple readers, and one of the readers is slow
(or dead), the other readers should not be impacted. This means that if
the writer wishes to publish a sample, but there is not room in its
history for the new sample because there are too many unacked samples,
old samples should be discarded to make room. Technically, it should
only discard samples if there are multiple readers, and there are
samples that are acked by _some_ readers, but this change does not take
any of that into account; it just drops the oldest sample.

The case where this is expected to matter is when there is a local
subscriber and there is a subscriber on the ground. When the link is
spotty, the ground-based subscriber will likely fall behind, but the
local subscriber should continue to get full-rate samples.